+++ src/core/ds.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert";
 import { describe, it } from "node:test";
-import { setHasValue } from "./utils.ts";
+import { setHasValue } from "./ds.ts";
 
 describe("core/utils", () => {
 	describe("setHasValue", () => {
